refactor: Move PasswordToggle out of the folder
